- created test method for clean text - Created a new python file for pre processing test methods - Renamed test_main to test_controllers - Uplaoded a testfiles folder
