+++ Scripts/app.py
@@ -1,6 +1,11 @@
 from flask import Flask
 from flask_cors import CORS
 from flask_jsonpify import jsonify
+from flask import g
+import pickle
+from flask import Flask
+from flask import request
+
 
 app = Flask(__name__)
 CORS(app)
@@ -19,6 +24,9 @@ def get_documents():
 @app.route('/health')
 def get_health():
     return jsonify({"status": "UP"})
+
+
+
 
 
 if __name__ == '__main__':
