+++ frontend/angular.json
@@ -123,6 +123,10 @@
           }
         }
       }
-    }},
-  "defaultProject": "frontend"
+    }
+  },
+  "defaultProject": "frontend",
+  "cli": {
+    "analytics": "09e7cb09-a792-4d44-afc2-4f554c89ba44"
+  }
 }
