+++ frontend/src/index.html
@@ -3,6 +3,10 @@
 <head>
   <meta charset="utf-8">
   <title>Frontend</title>
+
+
+
+
   <base href="/">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="icon" type="image/x-icon" href="favicon.ico">
