+++ tests/testfiles/100-02.txt
@@ -1,0 +1,93 @@
+
+
+
+Record date: 2108-03-14
+
+                     CAMPBELL EMERGENCY DEPT VISIT
+
+ 
+
+VALDEZ,HARLAN,JR.   845-41-54-4                VISIT DATE: 03/14/08
+
+The patient was seen and examined in the emergency department.  The 
+
+patient was seen by the Emergency Medicine resident.  I have 
+
+discussed the management with the resident.  I have also seen the 
+
+patient primarily and reviewed the medical record.  This is a brief 
+
+addendum to the medical record.
+
+HISTORY OF PRESENTING COMPLAINT:  Briefly, this is a 45-year-old 
+
+male who complains of several days of nausea, vomiting, and left 
+
+lower quadrant discomfort.  He also describes intermittent chest 
+
+pain, which he has had for a number of months without significant 
+
+change.  He was sent in from his primary care doctor today with 
+
+this pain and was also noted to have some EKG changes.  The patient 
+
+has no chest pain at the time of evaluation in the emergency 
+
+department and no shortness of breath.
+
+REVIEW OF SYSTEMS:  As indicated and otherwise negative.
+
+PAST MEDICAL HISTORY:  As indicated in the chart.
+
+SOCIAL HISTORY AND FAMILY HISTORY:  As indicated in the chart.
+
+PHYSICAL EXAMINATION:  On physical examination, the patient is very 
+
+well-appearing, a smiling, very pleasant gentleman in no acute 
+
+distress.  The blood pressure is 119/90, the pulse 82, and the 
+
+temperature 97.9 degrees.  Normocephalic and atraumatic.  The chest 
+
+is clear to auscultation.  The heart has a regular rate and rhythm. 
+
+The abdomen is soft.  He has left lower quadrant tenderness.  He 
+
+also, of note on cardiovascular examination, has a soft murmur 
+
+which he says he has had since childhood.  The extremities are 
+
+normal.  The neurologic examination is non-focal. 
+
+THERAPY RENDERED/COURSE IN ED:  This is a gentleman with abdominal 
+
+pain who will receive a CAT scan to rule out diverticulitis.  He 
+
+has also had some non-specific ST changes on his EKG.  He is 
+
+pain-free at this time.  He does not describe a classic exertional 
+
+pattern for his chest pain, but given that he is a diabetic and 
+
+with EKG changes, he will also be admitted for rule out MI.  A CT 
+
+is pending at the time of this dictation.
+
+DISPOSITION (including condition upon discharge):  As above.  The 
+
+patient's condition is currently stable. 
+
+___________________________________                    CK498/89095 
+
+JAY CARROLL, M.D.  JC72                                 D:03/14/08 
+
+                                                       T:03/14/08 
+
+Dictated by:  JAY CARROLL, M.D.  JC72 
+
+ ******** Not reviewed by Attending Physician ********
+
+
+
+
+
