+++ tests/testfiles/100-04.txt
@@ -1,0 +1,153 @@
+
+
+
+Record date: 2111-10-10
+
+    
+
+
+
+CCU JAR Transfer Note
+
+Admission Date: 10/8/11
+
+Transfer: 10/10/11
+
+Patient Name: Valdez, Harlan 
+
+MRN#: 7672624
+
+Cardiologist:  Dr. Nutter
+
+PCP:  Vicente Barker
+
+
+
+CC: Chest Pain    Cath    VF arrest    RCA stenting
+
+
+
+History of Present Illness (obtained on admission):  Pt is a 48 yo male with h/o DMII, hypercholesterolemia, Bipolar d/o, and depression who began to have sub-sternal day prior to admission in car and pre-syncope + profound weakness.  This CP was minimal, but the weakness made him pull over.  He had a repeat of these symptoms day of admission.  His EKG c/w 2/2107 showed flattened T-wave in V2 and TWI in V3 and flattened T-waves in I, aVL. His trop was negative, but MB index was elevated.  Due to T-wave flattening, history and elevated index it was decided to start on heparin and ASA and take to cath lab.  
+
+
+
+Cath showed right dominant system with prox Cx 40%, LAD clear, RCA prox 70-80% lesion and ostial PDA 90%.  During final dye injection, pt had VF arrest and 2 shocks.  Pt regained puls and was in AF (new) with RVR. Pt was started on amio.  Pt then began to experience discomfort in the RVR and it was decided to intervene.  POBA was done to ostial PDA.  A first no-eluting stent was placed in prox RCA and pt had dissection and thus 2cd stent was placed.  On admission to CCU, pt still in AF with RVR (120's).  He was on amio drip, BB, loaded on plavix, ASA, lipitor, integrilin and was placed on Avandia study.  His complaint of some mild chest pain (not same as anginal pain day before) thought to be from defibrillation.
+
+
+
+Past Medical History: DMII, hyperchol, bipolar,HTN, depression (s/p ECT)
+
+ 
+
+Medications on admission: ASA, Lipitor 20, lopressor 50 bid, folate, norvasc 5 qd; lithium, 300 bid; depakote 500 bid; sonata 10 mg qhs, doxylamine 25 qhs, mirtazapine 45 qd 
+
+
+
+Meds on Transfer: please see green sheets
+
+ 
+
+Medications: ASA, Lipitor 20, lopressor 50 bid, folate, norvasc 5 qd; lithium, 300 bid; depakote 500 bid; sonata 10 mg qhs, doxylamine 25 qhs, mirtazapine 45 qd 
+
+
+
+Allergies:  NKDA
+
+
+
+Family History:  family h/o CAD
+
+
+
+Social History: No EtOH, no tob, no illicits
+
+
+
+Review of Systems:  per HPI
+
+
+
+Allergies:  NKDA
+
+
+
+Family History:  family h/o CAD
+
+
+
+Social History: No EtOH, no tob, no illicits
+
+
+
+Review of Systems:  per HPI
+
+
+
+CCU course + plan:
+
+1)	Cards
+
+a.	Rhythm - on night of admission patient was started on an esmolol drip as well as amio bloused and rhythm converted to NSR.  Esmolol drip as well as amio was stopped and BB was escalated and patient has remained in NSR.
+
+i.	Ramp up lopressor as tolerated by BP
+
+b.	Pump - patient has remained euvolemic and had a Echo with EF 84% and aortic stenosis
+
+c.	Ischemia - was stented x 2 to the prox RCA lesion and was on integrilin x 24hrs prior.  He was started on plavix.
+
+i.	Cont plavix, lopressor, lisinopril, lipitor, ASA
+
+
+
+2)	Psych - patient with long history of bipolar disorder + depression.  He was on depakote, lithium and remeron as outpt.  He was seen by psychiatry here.
+
+a.	Continue depakote and lithium (had subthereapeutic level that psych thought was likely  due to non-compliance.
+
+b.	Continue remeron qhs
+
+c.	F/u TSH
+
+
+
+3)	DM - Blood sugars were originally elevated as amio drip he was originally on for AF contained dextrose. He has remained on NPH with RISS.
+
+a.	NPH, RISS
+
+
+
+4)	Prophy - Fragmin nexium
+
+
+
+
+
+LABS + PE - see today's progress note
+
+
+
+
+
+EKG - AFIB with RVR, diffuse t-wave flattening
+
+
+
+Impression:  48 yo male with h/o DMII, hypercholesterolemia, Bipolar d/o, and depression and CAD p/w CP and pre-sycope found to have RCA prox 70-80% lesion and ostial PDA 90% (stents to RCA and POBA to PDA).  Cath c/b VF arrest after dye load and resultant afib with RVR.
+
+
+
+Plan:  
+
+As outlined in CCU course.
+
+
+
+
+
+Victor Shepard MD
+
+39693
+
+   
+
+
+
